refactored sensor push activity
moved from structure modifying interface to bus interface
extending an addressable model item with a ->pushSensor()
method

+++ src/ds485/dsdevicebusinterface.cpp
@@ -157,7 +157,7 @@ namespace dss {
                                           _sensorIndex, kDSM_API_TIMEOUT,
                                           &retVal);
     DSBusInterface::checkResultCode(ret);
-    _device.setSensorValue(_sensorIndex, retVal);
+    _device.setSensorValue(_sensorIndex, (const unsigned int) retVal);
     return retVal;
   } // getSensorValue
